requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sse-starlette>=1.6.0

# HTTP client
httpx>=0.25.0

# Environment variables
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0

# Google Cloud Vertex AI
google-cloud-aiplatform[agent_engines]>=1.38.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
